refactor: review dirs hierarchy
