+++ packages/web/src/new/scenes/Main/Auth/Subviews/Restore/WalletPlaceholder.tsx
