--- packages/web/src/new/scenes/Main/Auth/Subviews/components/index.ts
@@ -1,1 +1,0 @@
-export { Checkbox } from './Checkbox';
